apply translations after everything else
