+++ app/templates/views/notifications/notification.html
@@ -75,7 +75,7 @@
     {% endif %}
 
     <div class="{{ 'letter-sent' if template.template_type == 'letter' else '' }}">
-      {{ template.html|safe }}
+      {{ template|string|translate_preview_template }}
     </div>
 
     {% if template.template_type == 'letter' %}
