+++ app/templates/views/send.html
@@ -45,7 +45,7 @@ template_id=template.id) ) }}
 <h2 class="heading-medium">
   {{ _('Your file will populate this template')}} ({{ template.name }})
 </h2>
-{{ template.html|safe }}
+{{ template|string|translate_preview_template }}
 
 <script>
   function isIE() {
