+++ app/templates/views/templates/_template_history.html
@@ -12,5 +12,5 @@
 </div>
 
 <div class="column-whole">
-  {{ template.html|safe }}
+  {{ template|string|translate_preview_template }}
 </div>
